+++ packages/core/src/abap/3_structures/structures/section_section.ts
